#208: Support multiple profile folders, configurable via command option
